+++ src/Compute/Compute/ChangeLog.md
@@ -19,9 +19,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added 'AssignedHost' field to Get-AzVM's instance view. The field shows the resource id of the virtual machine instance.
 * Added '-EncryptionAtHost' parameter to New-AzVm, New-AzVmss, New-AzVMConfig, New-AzVmssConfig, Update-AzVM, and Update-AzVmss
 * Added 'SecurityProfile' to Get-AzVM and Get-AzVmss return object
 * Added the '-InstanceView' switch as optional parameter to Get-AzHostGroup
+* Added the '-HostGroupId' parameter to New-AzVm.
 
 ## Version 4.2.1
 * Added warning when using `New-AzVmss` without "latest" image version